+++ twilio/http/http_client.py
@@ -1,6 +1,6 @@
 from requests import Request, Session
 
-from twilio.http import HttpClient, get_cert_file
+from twilio.http import HttpClient
 from twilio.http.response import Response
 
 
@@ -8,12 +8,8 @@ class TwilioHttpClient(HttpClient):
     """
     General purpose HTTP Client for interacting with the Twilio API
     """
-    def __init__(self, connection_pool=True):
-        if connection_pool:
-            self.session = Session()
-            self.session.verify = get_cert_file()
-        else: 
-            self.session = None
+    def __init__(self, pool_connections=True):
+        self.session = Session() if pool_connections else None
             
     def request(self, method, url, params=None, data=None, headers=None, auth=None, timeout=None,
                 allow_redirects=False):
@@ -33,11 +29,7 @@ class TwilioHttpClient(HttpClient):
         :return: An http response
         :rtype: A :class:`Response <twilio.rest.http.response.Response>` object
         """
-        session = self.session
-        if session is None:
-            session = Session()
-            session.verify = get_cert_file()
-
+        session = self.session or Session()
         request = Request(method.upper(), url, params=params, data=data, headers=headers, auth=auth)
 
         prepped_request = session.prepare_request(request)